Add csonpath JSON library to README
Hello, 
I've made a JSONPath library that can be use a dual-file library.
If used like so, the JSON is abstract, and the user need to define a few macros to use it. (and include csonpath.h)
Otherwise, the lib contains a json-c, yyjson and a python module implementations, but in this case it would be a 3 files libs.
The lib was made so it can be use either directly from one of the implementations, or as a headers-only lib.

Also, I've added it in JSON, but maybe a JSONPath section could be added.